- BUG: Some PHP bugs/typos for older versions
git-svn-id: file:///Users/jan/Dropbox/svn/FlexyAdmin/trunk@3566 0c02b2c5-a8bd-4e35-85b2-cf8cacb96951

Former-commit-id: 75ce8143b90bd76b4becbbeb9338a756b059ae9b
Former-commit-id: 351abd3c17de9759a9608ec1205f909c964bf923
Former-commit-id: c9b85648ee1d424c46ac05270581ceca207ad133 [formerly 69d71536ed1d9006cf97795cade5cb34ba150e94]
Former-commit-id: 381bb0ad09f7e3a990065a8fb22afa425e11e2ab

+++ sys/flexyadmin/libraries/plugins/Plugin_safe_assets.php
@@ -93,8 +93,11 @@ class Plugin_safe_assets extends Plugin {
   function _admin_logout() {
 		$logout=true;
     // Cleanup captcha's
-    foreach( glob('site/assets/_thumbcache/captcha*') as $file ) {
-      unlink($file);
+    $files = glob('site/assets/_thumbcache/captcha*');
+    if (is_array($files)) {
+      foreach( $files as $file ) {
+        unlink($file);
+      }
     }
 		$this->_safe_and_clean_all();
     if (!empty($this->removed)) return $this->_show();
